Moved loaders to Baseviewcontroller
